feature: MongoDB Integrtaion  CR:suprusty SANITY:Yes

+++ src/main/java/com/cisco/blog/persistence/service/APersistenceService.java
@@ -2,7 +2,7 @@ package com.cisco.blog.persistence.service;
 
 import java.util.List;
 
-import com.cisco.blog.data.ABlog;
+import com.cisco.blog.data.profile.ABlog;
 import com.cisco.blog.exception.persistence.TransactionException;
 
 
@@ -10,7 +10,7 @@ public abstract class APersistenceService
 {
 	abstract void addObject (ABlog aBlog) throws TransactionException;
 	abstract void updateObject (ABlog aBlog) throws TransactionException;
-	abstract ABlog getObjectById (Integer id,Class entityClass) throws TransactionException;
+	abstract ABlog getObjectById (Long id,Class entityClass) throws TransactionException;
 	abstract List<?> getListOfObject(Class entityClass) throws TransactionException;
 	abstract void deleteObject(Integer id,Class entityClass) throws TransactionException;
 	abstract Long getCountByBusinessAndNameText(ABlog aBlog,String namedQuery) throws TransactionException;
